refactor: move settings package from `no-meteor` to `node-app`
