+++ imports/node-app/core-services/settings/index.js
@@ -1,8 +1,8 @@
-import mutations from "./mutations";
-import queries from "./queries";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
-import { registerPluginHandler } from "./util/settingsConfig";
+import mutations from "./mutations/index.js";
+import queries from "./queries/index.js";
+import resolvers from "./resolvers/index.js";
+import schemas from "./schemas/index.js";
+import { registerPluginHandler } from "./util/settingsConfig.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
