+++ imports/node-app/core-services/settings/mutations/index.js
@@ -1,0 +1,5 @@
+import updateAppSettings from "./updateAppSettings.js";
+
+export default {
+  updateAppSettings
+};
