+++ imports/node-app/core-services/settings/queries/index.js
@@ -1,0 +1,5 @@
+import appSettings from "./appSettings.js";
+
+export default {
+  appSettings
+};
