+++ imports/node-app/core-services/settings/resolvers/Mutation/index.js
@@ -1,0 +1,7 @@
+import updateGlobalSettings from "./updateGlobalSettings.js";
+import updateShopSettings from "./updateShopSettings.js";
+
+export default {
+  updateGlobalSettings,
+  updateShopSettings
+};
