+++ imports/node-app/core-services/settings/resolvers/Query/index.js
@@ -1,0 +1,7 @@
+import globalSettings from "./globalSettings.js";
+import shopSettings from "./shopSettings.js";
+
+export default {
+  globalSettings,
+  shopSettings
+};
