--- imports/plugins/core/settings/server/mutations/index.js
@@ -1,5 +1,0 @@
-import updateAppSettings from "./updateAppSettings";
-
-export default {
-  updateAppSettings
-};
