--- imports/plugins/core/settings/server/queries/index.js
@@ -1,5 +1,0 @@
-import appSettings from "./appSettings";
-
-export default {
-  appSettings
-};
