--- imports/plugins/core/settings/server/resolvers/Mutation/index.js
@@ -1,7 +1,0 @@
-import updateGlobalSettings from "./updateGlobalSettings";
-import updateShopSettings from "./updateShopSettings";
-
-export default {
-  updateGlobalSettings,
-  updateShopSettings
-};
