--- imports/plugins/core/settings/server/resolvers/Query/index.js
@@ -1,7 +1,0 @@
-import globalSettings from "./globalSettings";
-import shopSettings from "./shopSettings";
-
-export default {
-  globalSettings,
-  shopSettings
-};
