--- imports/plugins/core/settings/server/resolvers/index.js
@@ -1,7 +1,0 @@
-import Mutation from "./Mutation";
-import Query from "./Query";
-
-export default {
-  Mutation,
-  Query
-};
